backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
supabase==2.3.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
email-validator==2.1.0
